Grade prediction and db storage
Changed the grade prediction algorithm to a Bayesian linear regression.
Store the data needed for the diagram in the db (instance of User). If
no new grades has been entered in Canvas, the tool will directly get
the data from the database.
Changed the way the app calculates the current average grade of the
students.